requirements.txt
asgiref==3.11.0
attrs==25.4.0
bcrypt==5.0.0
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.6.0
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
python-docx==1.2.0
docxtpl==0.20.2
jinja2==3.1.6
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
gunicorn==21.2.0
whitenoise==6.6.0
mammoth==1.6.0
python-dotenv==1.0.0
Pillow==10.2.0
